refactor(pine-go): move bench_cpu and bench_sleep to operators/transform
These two operators (transform_bench_cpu, transform_bench_sleep) are
built-in benchmark utilities that don't conflict with downstream code.
They belong in the always-imported transform package, not the
pine_bench-gated bench package.
