+++ pine-go/operators/transform/bench_sleep.go
@@ -1,4 +1,4 @@
-package bench
+package transform
 
 import (
 	"context"
